+++ Functions/TestDrive.ps1
@@ -17,13 +17,13 @@ function New-TestDrive ([Switch]$PassThru, [string] $Path) {
     $DriveName = "TestDrive"
 
     #setup the test drive
-    if ( -not (Get-PSDrive -Name $DriveName -ErrorAction SilentlyContinue) )
+    if ( -not (Test-Path "${DriveName}:\") )
     {
         New-PSDrive -Name $DriveName -PSProvider FileSystem -Root $directory -Scope Global -Description "Pester test drive" | Out-Null
     }
 
     #publish the global TestDrive variable used in few places within the module
-    if (-not (Get-Variable -Name $DriveName -Scope Global -ErrorAction SilentlyContinue))
+    if (-not (Test-Path "Variable:Global:DriveName"))
     {
         New-Variable -Name $DriveName -Scope Global -Value $directory
     }
@@ -72,7 +72,7 @@ function Get-TestDriveChildItem {
 function Remove-TestDrive {
 
     $DriveName = "TestDrive"
-    $Drive = Get-PSDrive -Name $DriveName -ErrorAction SilentlyContinue
+    $Drive = Get-PSDrive -Name $DriveName -ErrorAction (Get-IgnoreErrorPreference)
     $Path = ($Drive).Root
 
 
@@ -84,7 +84,7 @@ function Remove-TestDrive {
 
     if ( $Drive )
     {
-        $Drive | Remove-PSDrive -Force -ErrorAction SilentlyContinue
+        $Drive | Remove-PSDrive -Force -ErrorAction (Get-IgnoreErrorPreference)
     }
 
     if (Microsoft.PowerShell.Management\Test-Path -Path $Path)
@@ -92,7 +92,7 @@ function Remove-TestDrive {
         Microsoft.PowerShell.Management\Remove-Item -Path $Path -Force -Recurse
     }
 
-    if (Get-Variable -Name $DriveName -Scope Global -ErrorAction SilentlyContinue) {
+    if (Get-Variable -Name $DriveName -Scope Global -ErrorAction (Get-IgnoreErrorPreference)) {
         Remove-Variable -Scope Global -Name $DriveName -Force
     }
 }